Simplify integration instructions
The @aar syntax is only required if there also is a jar artifact to
distinguish from. It also has the side-effect to trigger "artifact only"
mode for Gradle, resulting in transitive module dependencies not being
resolved automatically, which was previously worked around by eitehr
specifying "transitive = true" or "com.google.zxing:core:3.2.1" directly.
neitehr is required if we just omit the @aar, which works fine as there is
no competing jar artifact.