Correct minSdkVersion to 19
zxing:core never increased the minimum version to 24. It increased it to
19: https://github.com/zxing/zxing/blob/44267115/android/AndroidManifest.xml#L34

This most recent bump happened here: https://github.com/zxing/zxing/commit/04595508b

Now, this disagrees with the release notes of 3.4.0:
https://github.com/zxing/zxing/releases/tag/zxing-3.4.0 which say the
minimum verison is API 24 due to Java 8. But these days, AGP desugars
Java 8 automatically, so you can, in fact, use it with older APIs.

See this for more information:
https://developer.android.com/studio/releases#4-0-0-desugar